+++ controllers/restore_test.go
@@ -19,6 +19,7 @@ package controllers
 
 import (
 	"context"
+	"errors"
 	"path/filepath"
 	"testing"
 	"time"
@@ -29,8 +30,13 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/dynamic"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 )
 
@@ -1936,6 +1942,188 @@ func Test_actLabelNotOnManagedClsRestore(t *testing.T) {
 		}
 	}
 	// clean up
+	if err := testEnv.Stop(); err != nil {
+		t.Fatalf("Error stopping testenv: %s", err.Error())
+	}
+}
+
+//nolint:funlen
+func TestRestoreReconciler_finalizeRestore(t *testing.T) {
+
+	testEnv := &envtest.Environment{
+		CRDDirectoryPaths: []string{
+			filepath.Join("..", "config", "crd", "bases"),
+			filepath.Join("..", "hack", "crds"),
+		},
+		ErrorIfCRDPathMissing: true,
+	}
+
+	ns1 := *createNamespace("backup-ns")
+	acmRestore1 := *createACMRestore("acm-restore", "backup-ns").
+		cleanupBeforeRestore(v1beta1.CleanupTypeNone).syncRestoreWithNewBackups(true).
+		veleroManagedClustersBackupName(latestBackupStr).
+		veleroCredentialsBackupName(latestBackupStr).
+		veleroResourcesBackupName(latestBackupStr).object
+
+	veleroRestoreFinalizer := *createRestore("velero-res", ns1.Name).
+		backupName("backup").
+		setFinalizer([]string{restoreFinalizer}).
+		object
+	veleroRestoreFinalizerDel := *createRestore("velero-res-terminate", ns1.Name).
+		backupName("backup").
+		setFinalizer([]string{restoreFinalizer}).
+		setDeleteTimestamp(metav1.NewTime(time.Now())).
+		object
+
+	scheme1 := runtime.NewScheme()
+	e1 := corev1.AddToScheme(scheme1)
+	e2 := veleroapi.AddToScheme(scheme1)
+	e3 := v1beta1.AddToScheme(scheme1)
+	if err := errors.Join(e1, e2, e3); err != nil {
+		t.Fatalf("Error adding apis to scheme: %s", err.Error())
+	}
+
+	cfg, err := testEnv.Start()
+	if err != nil {
+		t.Fatalf("Error starting testEnv: %s", err.Error())
+	}
+	k8sClient1, err := client.New(cfg, client.Options{Scheme: scheme1})
+	if err != nil {
+		t.Fatalf("Error starting client: %s", err.Error())
+	}
+	errs := []error{}
+	errs = append(errs, k8sClient1.Create(context.Background(), &ns1))
+	errs = append(errs, k8sClient1.Create(context.Background(), &acmRestore1))
+	errs = append(errs, k8sClient1.Create(context.Background(), &veleroRestoreFinalizer))
+	errs = append(errs, k8sClient1.Create(context.Background(), &veleroRestoreFinalizerDel))
+	if err := errors.Join(errs...); err != nil {
+		t.Fatalf("Error creating objects for test setup: %s", err.Error())
+	}
+
+	type fields struct {
+		Client          client.Client
+		KubeClient      kubernetes.Interface
+		DiscoveryClient discovery.DiscoveryInterface
+		DynamicClient   dynamic.Interface
+		Scheme          *runtime.Scheme
+		Recorder        record.EventRecorder
+	}
+	type args struct {
+		ctx               context.Context
+		c                 client.Client
+		acmRestore        *v1beta1.Restore
+		veleroRestoreList veleroapi.RestoreList
+	}
+	tests := []struct {
+		name    string
+		fields  fields
+		args    args
+		wantErr bool
+		errMsg  string
+	}{
+		{
+			name: "no restores, return nil",
+			args: args{
+				ctx:               context.Background(),
+				c:                 k8sClient1,
+				acmRestore:        &acmRestore1,
+				veleroRestoreList: veleroapi.RestoreList{},
+			},
+			wantErr: false,
+			errMsg:  "",
+		},
+		{
+			name: "has velero restores, but not marked for deletion, finalizer should NOT be removed",
+			args: args{
+				ctx:        context.Background(),
+				c:          k8sClient1,
+				acmRestore: &acmRestore1,
+				veleroRestoreList: veleroapi.RestoreList{
+					Items: []veleroapi.Restore{
+						veleroRestoreFinalizer,
+					},
+				},
+			},
+			wantErr: true,
+			errMsg:  "waiting for velero restores to be terminated",
+		},
+		{
+			name: "has velero restores, marked for deletion, but resource not found",
+			args: args{
+				ctx:        context.Background(),
+				c:          k8sClient1,
+				acmRestore: &acmRestore1,
+				veleroRestoreList: veleroapi.RestoreList{
+					Items: []veleroapi.Restore{
+						*createRestore("velero-1", ns1.Name).
+							backupName("backup").
+							setFinalizer([]string{restoreFinalizer}).
+							setDeleteTimestamp(metav1.NewTime(time.Now())).
+							object,
+					},
+				},
+			},
+			wantErr: true,
+			errMsg:  `restores.velero.io "velero-1" not found`,
+		},
+		{
+			name: "has velero restores, not marked for deletion and resource not found, delete must fail",
+			args: args{
+				ctx:        context.Background(),
+				c:          k8sClient1,
+				acmRestore: &acmRestore1,
+				veleroRestoreList: veleroapi.RestoreList{
+					Items: []veleroapi.Restore{
+						*createRestore("velero-1", ns1.Name).
+							backupName("backup").
+							setFinalizer([]string{restoreFinalizer}).
+							object,
+					},
+				},
+			},
+			wantErr: true,
+			errMsg:  `restores.velero.io "velero-1" not found`,
+		},
+		{
+			name: "has velero restores, marked for deletion, finalizer should be removed",
+			args: args{
+				ctx:        context.Background(),
+				c:          k8sClient1,
+				acmRestore: &acmRestore1,
+				veleroRestoreList: veleroapi.RestoreList{
+					Items: []veleroapi.Restore{
+						veleroRestoreFinalizerDel,
+					},
+				},
+			},
+			wantErr: true,
+			errMsg:  "waiting for velero restores to be terminated",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+
+			err := finalizeRestore(tt.args.ctx, tt.args.c, tt.args.acmRestore, tt.args.veleroRestoreList)
+
+			if err != nil && err.Error() != tt.errMsg {
+				t.Errorf("got an error but not the one expected error = %v, wantErr %v", err.Error(), tt.errMsg)
+			}
+
+			if (err != nil) != tt.wantErr {
+				t.Errorf("RestoreReconciler.finalizeRestore() error = %v, wantErr %v", err, tt.wantErr)
+			}
+
+			for _, veleroRestore := range tt.args.veleroRestoreList.Items {
+				if controllerutil.ContainsFinalizer(&veleroRestore, restoreFinalizer) &&
+					veleroRestore.GetDeletionTimestamp() != nil &&
+					veleroRestore.Name != "velero-1" {
+					t.Errorf("Velero restore marked for deletion but finalizer not removed name=%v",
+						veleroRestore.Name)
+
+				}
+			}
+		})
+	}
 	if err := testEnv.Stop(); err != nil {
 		t.Fatalf("Error stopping testenv: %s", err.Error())
 	}
